Rename 1.5 to 1.6 to match respective release number
