+++ contracts-legacy/v1.6.0/BaseModule.sol
@@ -39,7 +39,8 @@ contract BaseModule is Module {
      * @dev Throws if the wallet is locked.
      */
     modifier onlyWhenUnlocked(BaseWallet _wallet) {
-        verifyUnlocked(_wallet);
+        // solium-disable-next-line security/no-block-members
+        require(!guardianStorage.isLocked(_wallet), "BM: wallet must be unlocked");
         _;
     }
 
@@ -64,8 +65,7 @@ contract BaseModule is Module {
      * @dev Throws if the sender is not the owner of the target wallet or the module itself.
      */
     modifier onlyWalletOwner(BaseWallet _wallet) {
-        // Wrapping in an internal method reduces deployment cost by avoiding duplication of inlined code
-        verifyWalletOwner(_wallet);
+        require(msg.sender == address(this) || isOwner(_wallet, msg.sender), "BM: must be an owner for the wallet");
         _;
     }
 
@@ -103,24 +103,7 @@ contract BaseModule is Module {
     */
     function recoverToken(address _token) external {
         uint total = ERC20(_token).balanceOf(address(this));
-        bool success = ERC20(_token).transfer(address(registry), total);
-        require(success, "BM: recover token transfer failed");
-    }
-
-    /**
-     * @dev Verify that the wallet is unlocked.
-     * @param _wallet The target wallet.
-     */
-    function verifyUnlocked(BaseWallet _wallet) internal view {
-        require(!guardianStorage.isLocked(_wallet), "BM: wallet locked");
-    }
-
-     /**
-     * @dev Verify that the caller is the module or the wallet owner.
-     * @param _wallet The target wallet.
-     */
-    function verifyWalletOwner(BaseWallet _wallet) internal view {
-        require(msg.sender == address(this) || isOwner(_wallet, msg.sender), "BM: must be wallet owner");
+        ERC20(_token).transfer(address(registry), total);
     }
 
     /**
