+++ contracts-legacy/v1.6.0/GuardianUtils.sol
