+++ contracts-legacy/v1.6.0/LimitManager.sol
