+++ contracts-legacy/v1.6.0/SafeMath.sol
