+++ package.json
@@ -8,8 +8,8 @@
   "scripts": {
     "compile:lib": "npx etherlime compile --workingDirectory lib --solcVersion 0.5.4 --runs=200",
     "compile:legacy:1.3": "npx etherlime compile --workingDirectory contracts-legacy/v1.3.0 --buildDirectory build-legacy/v1.3.0 --solcVersion 0.5.4 --runs=999",
-    "compile:legacy:1.5": "npx etherlime compile --workingDirectory contracts-legacy/v1.5.0 --buildDirectory build-legacy/v1.5.0 --solcVersion 0.5.4 --runs=999",
-    "compile:legacy": "npm run compile:legacy:1.3 && npm run compile:legacy:1.5",
+    "compile:legacy:1.6": "npx etherlime compile --workingDirectory contracts-legacy/v1.6.0 --buildDirectory build-legacy/v1.6.0 --solcVersion 0.5.4 --runs=999",
+    "compile:legacy": "npm run compile:legacy:1.3 && npm run compile:legacy:1.6",
     "compile:infrastructure": "npx etherlime compile --workingDirectory contracts/infrastructure --solcVersion 0.5.4 --runs=999",
     "compile:modules": "npx etherlime compile --workingDirectory contracts/modules --runs=999",
     "compile:wallet": "npx etherlime compile --workingDirectory contracts/wallet --runs=999",
