+++ test/integration/makerV2Manager_loan.js
@@ -18,7 +18,7 @@ const Join = require("../../build/JoinLike");
 const Vat = require("../../build/VatLike");
 const FaucetUser = require("../../build/FaucetUser");
 const CdpManager = require("../../build/TestCdpManager");
-const MakerV1Manager = require("../../build-legacy/v1.5.0/MakerManager");
+const MakerV1Manager = require("../../build-legacy/v1.6.0/MakerManager");
 const MakerV2Manager = require("../../build/MakerV2Manager");
 const UpgradedMakerV2Manager = require("../../build/TestUpgradedMakerV2Manager");
 const TransferManager = require("../../build/TransferManager");
