+++ test/makerManager_loan.js
@@ -6,10 +6,10 @@ const {
 const { bigNumToBytes32, ETH_TOKEN } = require("../utils/utilities.js");
 const TestManager = require("../utils/test-manager");
 
-const LegacyWallet = require("../build-legacy/v1.5.0/BaseWallet");
-const Registry = require("../build-legacy/v1.5.0/ModuleRegistry");
-const GuardianStorage = require("../build-legacy/v1.5.0/GuardianStorage");
-const MakerManager = require("../build-legacy/v1.5.0/MakerManager");
+const LegacyWallet = require("../build-legacy/v1.6.0/BaseWallet");
+const Registry = require("../build-legacy/v1.6.0/ModuleRegistry");
+const GuardianStorage = require("../build-legacy/v1.6.0/GuardianStorage");
+const MakerManager = require("../build-legacy/v1.6.0/MakerManager");
 
 // Testing the Legacy MakerManager contract
 describe("Legacy MakerManager Module", function () {
