+++ test/makerV2Manager_loan.js
@@ -10,7 +10,7 @@ const { HashZero, AddressZero } = ethers.constants;
 const TestManager = require("../utils/test-manager");
 const GemJoin = require("../build/GemJoin");
 const Registry = require("../build/ModuleRegistry");
-const MakerV1Manager = require("../build-legacy/v1.5.0/MakerManager");
+const MakerV1Manager = require("../build-legacy/v1.6.0/MakerManager");
 const MakerV2Manager = require("../build/MakerV2Manager");
 const UpgradedMakerV2Manager = require("../build/TestUpgradedMakerV2Manager");
 const MakerRegistry = require("../build/MakerRegistry");
