+++ test/transferManager.js
@@ -7,7 +7,7 @@ const Registry = require("../build/ModuleRegistry");
 const TransferStorage = require("../build/TransferStorage");
 const GuardianStorage = require("../build/GuardianStorage");
 const TransferModule = require("../build/TransferManager");
-const LegacyTransferManager = require("../build-legacy/v1.5.0/TransferManager");
+const LegacyTransferManager = require("../build-legacy/v1.6.0/TransferManager");
 const TokenPriceProvider = require("../build/TokenPriceProvider");
 const ERC20 = require("../build/TestERC20");
 const TestContract = require("../build/TestContract");
